refactor(scripts): extract the export HTML and URL helpers into scripts/lib
`verify-export.mjs` carried its own HTML scraper and its own `package.json`
`homepage` basePath derivation. A second post-build script needs both, and two
copies of the basePath logic drift in exactly the place nobody tests: a
repository-site fork publishing under `/repo/`.

No behaviour change. `scripts/lib/html.mjs` holds the tag/attribute/meta
readers; `scripts/lib/site.mjs` holds `readSiteConfig` plus the route helpers,
returned as one bundle bound to the export directory so call sites keep their
current shape. `exportedFileExists` is now a thin wrapper over the shared
`exportFileFor`, which returns the resolved path so a size measurement can weigh
the file instead of only asking whether it exists.

Verified with `npm run verify-export` against a real `out/` (14 pages OK) and the
existing 17 script tests.
